pyproject.toml
[project]
name = "contract_analysis_flow_workshop"
version = "0.1.0"
description = "contract-analysis-flow-workshop using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.114.0,<1.0.0",
    "docling>=2.15.1",
    "weaviate-client>=4.13.2",
]

[project.scripts]
kickoff = "contract_analysis_flow_workshop.main:kickoff"
run_crew = "contract_analysis_flow_workshop.main:kickoff"
plot = "contract_analysis_flow_workshop.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
